TST: move epics binary calling utilities, server tests
